Moving onRoll to listener-misc-handler
